LayoutTests/media/media-can-load-when-hidden-expected.txt

Test HTMLMediaElement to be sure that the video is getting loaded even if the element is hidden.

EVENT(durationchange)
END OF TEST
